project/requirements.txt
aiofiles==0.6.0
celery==4.4.7
fastapi==0.64.0
flower==0.9.7
Jinja2==3.0.3
pytest==6.2.4
redis==3.5.3
requests==2.25.1
uvicorn==0.13.4
selenium==4.1.5
pandas==1.4.2
gspread==5.3.2
pillow==9.1.0
xlsxwriter==3.0.3